Alterado exclusão de notas para antes de excluir, zerar nota (para recalcular media) e para antes de excluir checar se a nota existe

+++ trunk/modules/Avaliacao/Views/DiarioAjaxController.php
@@ -264,9 +264,14 @@ class DiarioAjaxController extends Core_Controller_Page_EditController
         $this->setService() &&
         $this->validatesPresenceOfComponenteCurricularId(false))
     {
-      $this->getService()->deleteNota($this->getRequest()->componente_curricular_id, $this->getRequest()->etapa);
-      $this->saveService();
-      $this->appendMsg('Nota removida com sucesso.', 'notice');
+      if (! $this->getNotaAtual())
+        $this->appendMsg('Nota inexistente ou já removida.', 'notice');
+      else
+      {
+        $this->getService()->deleteNota($this->getRequest()->componente_curricular_id, $this->getRequest()->etapa);
+        $this->saveService();
+        $this->appendMsg('Nota removida com sucesso.', 'notice');
+      }
     }
   }
